% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{baro_spdf}
\alias{baro_spdf}
\title{smp_stats
baro_spdf}
\format{
Spatial Points Dataframe
}
\usage{
baro_spdf
}
\description{
A spatial dataset containing the latitude, longitude, smp id, and uid of the MARS barometric sensor network.
}
\keyword{datasets}
