% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{baroRasterPlot}
\alias{baroRasterPlot}
\title{Barometric Pressure Raster Plot}
\usage{
baroRasterPlot(baro)
}
\arguments{
\item{baro}{a dataframe with columns: \code{smp_id, baro_psi, day, year}}
}
\value{
p, a ggplot2 plot
}
\description{
Create a plot of barometric pressurs from each sensor
}
