% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_simulation_functions.R
\name{depth.volume}
\alias{depth.volume}
\alias{vol.to.depth}
\alias{depth.to.vol}
\title{Return Volume or Depth
Input maximum depth and maximum volume of SMP, along with depth or volume, to return the volume or depth.}
\usage{
vol.to.depth(maxdepth_ft, maxvol_cf, vol_cf)

\method{depth}{to.vol}(maxdepth_ft, maxvol_cf, depth_ft)
}
\arguments{
\item{maxdepth_ft}{depth of SMP when full, (ft)}

\item{maxvol_cf}{volume of SMP when full (cf)}

\item{vol_cf}{input or output volume (cf)}

\item{depth_ft}{input or output depth (cf)}
}
\value{
Output is either volume (cf) or depth (ft)
}
\description{
Return Volume or Depth
Input maximum depth and maximum volume of SMP, along with depth or volume, to return the volume or depth.
}
