% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetchRain}
\alias{fetchRain}
\title{Fetch rainfall data}
\usage{
fetchRain(conn, smp_id, source = c("gage", "radar"), start_date, end_date)
}
\arguments{
\item{conn}{Connection to database}

\item{smp_id}{SMP ID for the desired rainfall data}

\item{source}{Source of the rainfall data: Gage or Radar}

\item{start_date}{Start of the rainfall period (yyyy-mm-dd format)}

\item{end_date}{End of the rainfall period (yyyy-mm-dd format)}
}
\value{
Data frame of rainfall data between the start and end date
}
\description{
Fetch rainfall data
}
