% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{get_legend}
\alias{get_legend}
\title{Get Legend}
\usage{
get_legend(myggplot)
}
\arguments{
\item{myggplot}{ggplot object}
}
\value{
Returns a combined legend
}
\description{
Pull legend from separate ggplots for combined plot
}
