% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_functions.R
\name{lookupPrivateSMPs}
\alias{lookupPrivateSMPs}
\title{Return query results of private SMPs}
\usage{
lookupPrivateSMPs(con, tracking_numbers)
}
\arguments{
\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}

\item{tracking_number}{chr, tracking number for the requested SMP}
}
\value{
Output will be a dateframe with four columns:
  
    \item{tracking_number}{chr, requested SMP tracking number}
    \item{project name}{chr, SMP poject name}
    \item{smp_id}{int, SMP ID}
    \item{plan_label}{chr, name that the SMP is given on the plan set}
  
  If a tracking number is not found, that row will include an error message
  in the "project name" column, and NAs in "smp_id" and "plan_label".
}
\description{
Returns private SMP tracking number, project name, SMP ID, and plan label.
}
