% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsCheckSMPSnapshot}
\alias{marsCheckSMPSnapshot}
\title{Check if a data snapshot for an SMP exists}
\usage{
marsCheckSMPSnapshot(con, smp_id, ow_suffix, request_date)
}
\arguments{
\item{con}{An ODBC connection to the MARS Analysis database returned by odbc::dbConnect}

\item{smp_id}{vector of chr, SMP ID, where the user has requested data}

\item{ow_suffix}{vector of chr, OW Suffix, where the user has requested data}

\item{request_date}{single date or vector the length of SMP ID, either "YYYY-MM-DD" or "today", of the request data}
}
\value{
vector of boolean outcomes,
  
    \item{snapshot_present}{bool, TRUE or FALSE value based on whether a snapshot can exist or is made}
}
\description{
Returns a vector of boolean outcomes, TRUE if a snapshot exists or is written, FALSE if one does not exist
}
