% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_water_level_functions.R
\name{marsDraindownAssessment}
\alias{marsDraindownAssessment}
\title{Draindown Assessment}
\usage{
marsDraindownAssessment(
  level_ft,
  eventdepth_in,
  event_id_check,
  designdepth_in,
  storage_depth_ft,
  draindown_hr,
  subsurface = c(TRUE, FALSE)
)
}
\arguments{
\item{level_ft}{Observed water level data (ft)}

\item{eventdepth_in}{Rainfall event depth (in)}

\item{designdepth_in}{Depth that the SMP is designed to managed (in)}

\item{storage_depth_ft}{Maximum storage depth of system (ft)}

\item{draindown_hr}{Draindown calculated by marsDraindown_hr (hr)}

\item{subsurface}{If the SMP is subsurface, TRUE. If surface, FALSE.}
}
\value{
Output is a number based on draindown time and draindown time is rain event was scaled to the design depth. Outputs are codes for the following messages: 
 \describe{
       \item{\code{2}}{Draindown is below the target duration, and scaled draindown is below the target duration.}
       \item{\code{1}}{Draindown is above the target duration, but scaled draindown is below the target duration.}
       \item{\code{0}}{Rainfall occurs during recession period in specified depth range.}
       \item{\code{-1}}{Draindown is below the target duration, but scaled draindown is above the target duration.}
       \item{\code{-2}}{Draindown is above the target duration, and scaled draindown is above the target duration.}
 }
}
\description{
Assess Draindown Rate based on storm size
}
