% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_downloader_helper.R
\name{marsFetchBaroData}
\alias{marsFetchBaroData}
\title{Fetch barometric pressure data for a target SMP, date range, and interval}
\usage{
marsFetchBaroData(con, target_id, start_date, end_date,
  data_interval = c("5 mins", "15 mins"))
}
\arguments{
\item{con}{An ODBC connection to the MARS Analysis database returned by odbc::dbConnect}

\item{target_id}{chr, single SMP ID where the user has requested data}

\item{start_date}{POSIXct, format: "YYYY-MM-DD", start of data request range}

\item{end_date}{POSIXct, format: "YYYY-MM-DD", end of data request range}

\item{data_interval}{chr, \code{"5 mins"} or \code{"15 mins"}, interval at which baro data will be returned.}
}
\value{
Output will be a dataframe with four columns: "dtime_est" (POSIXct, format: "YYYY-MM-DD HH:MM:SS"),
   "baro_psi" (num), "smp_id" (chr), "neighbors" (int).
    If the target SMP has an on-site baro with data, the "neighbors" column will be NA.
    If there are fewer than five baros to interprolate from, based on \code{\link{marsInterpolateBaro}},
    all columns other than "dtime_est" will be NA.
}
\description{
Returns a data frame with datetime, barometric pressure, smp id, and number of neighbors
  interpolated from to collect the data.
}
\seealso{
\code{\link{marsInterpolateBaro}}
}
