% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsFetchLevelData}
\alias{marsFetchLevelData}
\title{Fetch water level data for an SMP}
\usage{
marsFetchLevelData(
  con,
  target_id,
  ow_suffix,
  start_date,
  end_date,
  sump_correct
)
}
\arguments{
\item{con}{An ODBC connection to the MARS Analysis database returned by odbc::dbConnect}

\item{target_id}{vector of chr, SMP ID, where the user has requested data}

\item{ow_suffix}{vector of chr, SMP ID, where the user has requested data}

\item{start_date}{string, format: "YYYY-MM-DD", start of data request range}

\item{end_date}{string, format: "YYYY-MM-DD", end of data request range}

\item{sump_correct}{logical, TRUE if water level should be corrected for to account for sump depth}
}
\value{
Output will be a dataframe with the following columns: 

    \item{ow_leveldata_uid}{int}
    \item{dtime_est}{POSIXct datetime}
    \item{level_ft}{num, recorded water level in feet}
    \item{ow_uid}{num, ow_uid derived from smp_id and ow_suffix}
}
\description{
Returns a data frame with requested SMP water level data
}
\seealso{
\code{\link{marsFetchRainfallData}}, \code{\link{marsFetchRainEventData}}, \code{\link{marsFetchMonitoringData}}
}
