% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsFetchMonitoringData}
\alias{marsFetchMonitoringData}
\title{Fetch monitoring data for an SMP}
\usage{
marsFetchMonitoringData(
  con,
  target_id,
  ow_suffix,
  start_date,
  end_date,
  sump_correct = TRUE,
  rain_events = TRUE,
  rainfall = TRUE,
  level = TRUE,
  daylight_savings = FALSE,
  debug = FALSE
)
}
\arguments{
\item{con}{An ODBC connection to the MARS Analysis database returned by odbc::dbConnect}

\item{target_id}{vector of chr, SMP ID, where the user has requested data}

\item{ow_suffix}{vector of chr, SMP ID, where the user has requested data}

\item{start_date}{string, format: "YYYY-MM-DD", start of data request range}

\item{end_date}{string, format: "YYYY-MM-DD", end of data request range}

\item{sump_correct}{logical, TRUE if water level should be corrected for to account for sump depth}

\item{rain_events}{logical, TRUE if rain event data should be included in result}

\item{rainfall}{logical, TRUE if rainfall data should be included in result}

\item{level}{logical, TRUE if water level should be included in result}

\item{daylight_savings}{logical, Adjust for daylight savings time? when doing QAQC this should be FALSE because the water level data does not spring forward}
}
\value{
Output will be a list consisting of a combination of the following:

    \item{Rain Event Data}{dataframe, output from \code{\link{marsFetchRainEventData}}}
    \item{Rain Gage Data}{dataframe, output from \code{\link{marsFetchRainGageData}}}
    \item{Level Data}{dataframe, output from \code{\link{marsFetchLevelData}}, plus rainfall_gage_event_uids}
}
\description{
Returns a list with data frames with rain data and rain event data from the rain gage closest to the request SMP, and level data
}
\seealso{
\code{\link{marsFetchRainGageData}}, \code{\link{marsFetchLevelData}}, \code{\link{marsFetchRainEventData}}
}
