% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars-gap_fill_and_rain_gage.R
\name{marsFetchRainGageData}
\alias{marsFetchRainGageData}
\title{Return a dataframe with rain gage data}
\usage{
marsFetchRainGageData(con, target_id, start_date, end_date,
  daylightsavings)
}
\arguments{
\item{con}{Formal class 'PostgreSQL', a connection to the MARS Analysis database}

\item{target_id}{chr, an SMP_ID that where the user has requested data}

\item{start_date}{POSIXct, format: "YYYY-MM-DD", start of data request range}

\item{end_date}{POSIXct, format: "YYYY-MM-DD", end of data request range}

\item{daylight_savings}{logi, Adjust for daylight savings time? when doing QAQC
this should be \code{FALSE} because the water level data does not spring forwards.}
}
\value{
Output will be a data frame with four columns: dtime_est (POSIXct datetime with tz = EST or EDT as specified by \code{daylight_savings}), rainfall_in,
  gagename, and event_id. This data will correspond to the specified SMP and date range.
}
\description{
Return data from the rain gage nearest a target SMP, for a specified date range.
}
\seealso{
\code{\link[pwdgsi]{gapFillEventID}}, \code{\link{detectEvents}}
}
