% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsFetchSMPSnapshot}
\alias{marsFetchSMPSnapshot}
\title{Fetch data snapshot for an SMP}
\usage{
marsFetchSMPSnapshot(con, smp_id, ow_suffix, request_date)
}
\arguments{
\item{con}{An ODBC connection to the MARS Analysis database returned by odbc::dbConnect}

\item{smp_id}{vector of chr, SMP ID, where the user has requested data}

\item{ow_suffix}{vector of chr, OW Suffix, where the user has requested data}

\item{request_date}{single date or vector the length of SMP ID, either "YYYY-MM-DD" or "today", of the request data}
}
\value{
Output will be a dataframe with the following columns: 
  
    \item{smp_id}{chr, requested SMP ID}
    \item{ow_suffix}{chr, request OW suffix}
    \item{ow_uid}{num, ow_uid derived from smp_id and ow_suffix} 
    \item{request_date}{chr, date of requested snapshot, format: "YYYY-MM-DD}
    \item{snapshot_uid}{num}
    \item{dcia_ft2}{num}
    \item{storage_footprint_ft2}{num}
    \item{orifice_diam_in}{num}
    \item{infil_footprint_ft2}{num}
    \item{assumption_orificeheight_ft}{num}
    \item{storage_depth_ft}{num}
    \item{sumpdepth_ft}{num}
    \item{lined}{chr, 0 or 1 for unlined or lined, respectively}
    \item{surface}{chr, 0 or 1 for subsurface or surface, respectively}
    
    If a requested smp_id and ow_suffix combination is not valid, the function will return NAs for that snapshot.
    Function queries the SQL function "insert_snapshot", which inserts rows into the MARS table "snapshot", and, in turn, "snapshot_metadata".
}
\description{
Returns a data frame with requested snapshot date, SMP ID, OW suffix, and SMP snapshot
}
