% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_functions.R
\name{marsGapFillEventID}
\alias{marsGapFillEventID}
\title{Return a dataset with event IDs for zero-punctuated timesteps.}
\usage{
marsGapFillEventID(event_low, event_high)
}
\arguments{
\item{event_low}{num, event ID of preceding rainfall.}

\item{event_high}{num, event ID of following rainfall.}
}
\value{
Output will be a vector containing either \code{NA} or \code{event low}. If one or both events are
\code{NA}, return {NA}. If the event IDs are not equal, return \code{NA}, since this is a boundary between
events. If the event IDs are equal, return \code{event low}.
}
\description{
Each rainfall event must be zero-punctuated. These zeroes are given event IDs based on the IDs of the
rainfall that precede and follow them, given by \code{\link{marsDetectEvents}}.
}
