% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsMetricsTable}
\alias{marsMetricsTable}
\title{Add metrics to an existing water level or combined plot}
\usage{
marsMetricsTable(
  in_plot,
  obs_RSPU = obs_RSPU,
  obs_infil_inhr = obs_infil_inhr,
  obs_draindown_hr = obs_draindown_hr,
  obs_overtopping = obs_overtopping,
  sim_RSPU = sim_RSPU,
  sim_infil_inhr = sim_infil_inhr,
  sim_draindown_hr = sim_draindown_hr,
  sim_overtopping = sim_overtopping
)
}
\arguments{
\item{in_plot}{ggplot object without annotative metrics table}

\item{obs_RSPU}{num, Metric: Observed relative percentage of storage used, see \code{marsPeakStorage_percent} (optional)}

\item{obs_infil_inhr}{num, Metric: Observed infiltration rate in inches per hour, see \code{marsInfiltrationRate_inhr} (optional)}

\item{obs_draindown_hr}{num, Metric: Observed draindown time in hours, see \code{marsDraindown_hr} (optional)}

\item{obs_overtopping}{bool, Metric: Observed overtopping boolean, see \code{marsOvertoppingCheck_bool} (optional)}

\item{sim_RSPU}{num, Metric: Simulated relative percentage of storage used, see \code{marsPeakStorage_percent} (optional)}

\item{sim_infil_inhr}{num, Metric: Simulated infiltration rate in inches per hour, see \code{marsInfiltrationRate_inhr} (optional)}

\item{sim_draindown_hr}{num, Metric: Simulated draindown time in hours, see \code{marsDraindown_hr} (optional)}

\item{sim_overtopping}{bool, Metric: Simulated overtopping boolean, see \code{marsOvertoppingCheck_bool} (optional)}
}
\value{
Output ggplot object adding metrics when necessary
}
\description{
Return the gpglot object, with the metrics added to the object as a tableGrob annotation
}
