% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsOvertoppingPlot}
\alias{marsOvertoppingPlot}
\title{Add metrics to an existing water level or combined plot}
\usage{
marsOvertoppingPlot(
  data,
  design_storm,
  event_dates = NULL,
  event_descriptions = NULL
)
}
\arguments{
\item{data}{dataframe, a data frame with a unique row each containing ow_uid, radar_event_uid, ow_suffix, eventdatastart_edt, smp_id, eventavgintensity_inhr, eventpeakintensity_inhr, eventdepth_in, and overtop}

\item{design_storm}{num, a numeric value fro the design stomr in inches, see \code{marsFetchSMPSnapshot}}

\item{event_dates}{date, a vector of dates to show up as major events in the monitoring locations history (maintenance, retrofit, etc.)}

\item{event_descriptions}{char, a vector of strings corresponding to labels fo each major event on the plot (eg, "pipe jetting", "filter bag replaced")}
}
\value{
Output ggplot object plotting overtopping, event peak intensity, and date
}
\description{
Return the gpglot object, with the metrics added to the object as a tableGrob annotation
}
