% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsRainfallPlot}
\alias{marsRainfallPlot}
\title{Plot hyetograph}
\usage{
marsRainfallPlot(dtime_est, rainfall_in, raingage, event,
  reverse_y = FALSE)
}
\arguments{
\item{dtime_est}{vector, POSIXct datetimes representing a single rain event}

\item{rainfall_in}{vector, num, rainfall in inches for that rain event}

\item{raingage}{chr, Label for the hyetograph for what rain gage the data came from}

\item{event}{chr, label for the hyetograph for what rain gage the data came from}

\item{reverse_y}{logical, whether the Y axes should be reversed}
}
\value{
Output will be a ggplot2 object of the hyetograph. Currently,
   the graphical parameters are hard coded and the legend is deleted.
   This may change in future versions.
}
\description{
Return hyetograph of events processed by \code{\link{marsDetectEvents}}
}
\examples{
gage_temp <- mutate(marsSampleRain, 
  event_id = marsDetectEvents(dtime_edt = marsSampleRain$dtime_edt, 
  rainfall_in = marsSampleRain$rainfall_in, 
  iet_hr = 6, mindepth_in = 0.10)) \%>\% filter(event_id == 2)
  
marsRainfallPlot(dtime_edt = gage_temp$dtime_edt, 
  rainfall_in = gage_temp$rainfall_in, raingage = 2, event = 2)   
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}
}
