% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{marsSampleBaro}
\alias{marsSampleBaro}
\title{Sample Baro Data}
\format{
A list containing a dataframe with 9 rows and 3 variables, and a character. 
\describe{
  \item{smp_id}{SMP ID}
  \item{baro_psi}{barometric pressure at the corresponding SMP}
  \item{weight}{weight assigned based on distance from target SMP}
  \item{target_ID}{SMP ID for which the user requests data}
}
}
\usage{
\code{marsSampleBaro}
}
\description{
A dataset containing barometric pressures and inverse distance weights for SMPs,
along with a target SMP ID, for which the user requests data.
}
\seealso{
\code{\link{marsInterpolateBaro}}
}
\keyword{datasets}
