% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{marsSampleBaro_plot}
\alias{marsSampleBaro_plot}
\title{Sample Baro Data for Plotting}
\format{
a dataframe with 2890 rows and 5 variables
\describe{
  \item{dtime_est}{POSIXct datetime, YYYY-MM-DD HH:MM:SS}
  \item{smp_id}{factor, SMP ID}
  \item{baro_psi}{num, barometric pressure (psi)}
  \item{neighbors}{int, number of nearby barometric sensors with recorded data}
  \item{weight}{num, weight of baro data based on distance, for interpolation}
  }
}
\usage{
marsBaroRasterPlot
}
\description{
A dataset containing datetime, smp_id, barometric pressure, number of neighbors, and weight
}
\keyword{datasets}
