% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{marsSampleRain}
\alias{marsSampleRain}
\title{Sample Rain Data}
\format{
A dataframe with 105 rows and 3 variables
\describe{
  \item{dtime_edt}{POSIXct datetime, YYYY-MM-DD HH:MM:SS}
  \item{rainfall_in}{rainfall total in the 15 minutes preceding the corresponding datetime}
  \item{gage_uid}{rain gage ID}
}
}
\usage{
marsSampleRain
}
\description{
A dataset containing datetimes, rainfall in inches, and rainfall gage id.
}
\seealso{
\code{\link{marsDetectEvents}}, \code{\link{marsRainfallPlot}}
}
\keyword{datasets}
