% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsSavePlot}
\alias{marsSavePlot}
\title{Wrapper function for ggsave that ensures text appears correctly for saved images of pwdgsi plots}
\usage{
marsSavePlot(
  in_plot = last_plot(),
  plot_type,
  filename,
  resolution = "med",
  ...
)
}
\arguments{
\item{in_plot}{ggplot object without annotative metrics table}

\item{plot_type}{text, one of 'combined', 'level', or 'rain' to determine the dimensions to use}

\item{filename}{text, file location to save the document refer to \code{ggsave}}

\item{resolution}{text, one of 'low', 'med', or 'high'. Each consists of default options that can be overwritten with ggsave arguments.}

\item{...}{overwrite assumed values of resolution. Accepts existing arguments to ggsave, see \code{ggsave} (optional)}
}
\value{
Output ggplot object adding metrics when necessary
}
\description{
Return the gpglot object, with the metrics added to the object as a tableGrob annotation
}
