% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsWriteInfiltrationData}
\alias{marsWriteInfiltrationData}
\title{Write Infiltration Performance Data to Database}
\usage{
marsWriteInfiltrationData(con, infiltration_rate_inhr, baseline_ft = NA,
  ow_uid, rainfall_gage_event_uid, snapshot_uid,
  observed_simulated_lookup_uid)
}
\arguments{
\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}

\item{infiltration_rate_inhr}{vector, numeric, infiltration rate (in/hr)}

\item{baseline_ft}{vector, numeric, point that water level returns to at the end of an event (default ??)}

\item{ow_uid}{vector, numeric observation well UID}

\item{rainfall_gage_event_uid}{vector, numeric}

\item{snapshot_uid}{vector, numeric}

\item{observed_simulated_lookup_uid}{vector, numeric, 1 if observed, 2 if simulated}
}
\value{
\code{TRUE} if the write is succesful, or an error message if unsuccessful
}
\description{
Receive vectors of infiltration rate and infiltration baseline data calculated with \code{\link{marsInfiltrationRate_inhr}},
gather data, and write to MARS Analysis Database performance_infiltration table. Replaces error codes with NAs and moves
them to a separate column, \code{error_lookup_uid}.
}
\examples{

marsWriteInfiltrationData(con = mars, 
  infiltration_rate_inhr = summary_250$infiltration_rate_inhr,
  ow_uid = summary_250$ow_uid,
  rainfall_gage_event_uid = summary_250$rainfall_gage_event_uid,
  snapshot_uid = summary_250$snapshot_uid,
  observed_simulated_lookup_uid = summary_250$observed_simulated_lookup_uid)

}
\seealso{
\code{\link[pwdgsi]{marsWriteOvertoppingData}}, \code{\link{marsWritePercentStorageData}}
}
