% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsWritePercentStorageData}
\alias{marsWritePercentStorageData}
\title{Write Percent of Storaged Used Data to Database}
\usage{
marsWritePercentStorageData(
  con,
  percentstorageused_peak,
  percentstorageused_relative,
  ow_uid,
  radar_event_uid,
  snapshot_uid,
  observed_simulated_lookup_uid
)
}
\arguments{
\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}

\item{percentstorageused_peak}{vector, numeric, peak percent of storage (\%)}

\item{percentstorageused_relative}{vector, numeric, relative percent storage (\%)}

\item{ow_uid}{vector, numeric observation well UID}

\item{radar_event_uid}{vector, numeric event UIDs for rain events from radar data}

\item{snapshot_uid}{vector, numeric}

\item{observed_simulated_lookup_uid}{vector, numeric, 1 if observed, 2 if simulated}
}
\value{
\code{TRUE} if the write is successful, or an error message if unsuccessful
}
\description{
Receive vectors of raw and relative percent storage data, calculated with \code{\link{marsPeakStorage_percent}},
gather data, and write to MARS Analysis performance_percentstorage table
}
\examples{

marsWritePercentStorageData(con = mars, 
   percentstorageused_peak = summary_250$percentstorageused_peak,
   percentstorageused_relative = summary_250$percentstorageused_relative,
   ow_uid = summary_250$ow_uid,
   radar_event_uid = summary_250$rainfall_gage_event_uid,
   snapshot_uid = summary_250$snapshot_uid,
   observed_simulated_lookup_uid = summary_250$observed_simulated_lookup_uid)

}
\seealso{
\code{\link{marsWriteOvertoppingData}},  \code{\link{marsWriteDraindownData}}
}
