% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_write_functions.R
\name{marsWriteSaturatedData}
\alias{marsWriteSaturatedData}
\title{Write Saturated Performance Data to Database}
\usage{
marsWriteSaturatedData(infiltration_rate_inhr, recession_rate_inhr, ow_uid,
  rainfall_gage_event_uid, snapshot_uid, observed_simulated_lookup_uid,
  con)
}
\arguments{
\item{infiltration_rate_inhr}{vector, numeric, infiltration rate (in/hr)}

\item{recession_rate_inhr}{vector, numeric, recession rate (in/hr)}

\item{ow_uid}{vector, numeric observation well UID}

\item{rainfall_gage_event_uid}{vector, numeric}

\item{snapshot_uid}{vector, numeric}

\item{observed_simulated_lookup_uid}{vector, numeric, 1 if observed, 2 if simulated}

\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}
}
\value{
\code{TRUE} if the write is succesful, or an error message if unsuccessful
}
\description{
Receive vectors of infiltration rate and recession rate, calculated with \code{\link{marsSaturatedPerformance_inhr}}
gather data, and write to MARS Analysis Database performance_saturated table
}
\seealso{
\code{\link[pwdgsi]{marsWriteOvertoppingData}}, \code{\link{marsWritePercentStorageData}}
}
