% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_rainfall_functions.R
\name{minInterval_hr}
\alias{minInterval_hr}
\title{Minimum Measurement Interval}
\usage{
minInterval_hr(dtime_est)
}
\arguments{
\item{dtime_est}{vector, POSIXct date times, in ascending order}
}
\value{
Minimum time measurement interval, in hours
}
\description{
Calculate minimum measurement interval for use in other functions.
}
