% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{monitoringdata_250}
\alias{monitoringdata_250}
\title{SMP Monitoring Data}
\format{a list with 3 dataframes}
\usage{
\code{monitoringdata_250}
}
\description{
A dataset containing water level, rainfall, and rain event data for SMP-250-1-1 from 2018-01-01 to 2018-03-01.
}
\keyword{datasets}
