% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{obs_250_all}
\alias{obs_250_all}
\title{Sample SMP Data}
\format{
a dataframe with 2890 rows and 5 variables
\describe{
 \item{dtime_est}{POSIXct datetime, YYYY-MM-DD HH:MM:SS}
 \item{level_ft}{num, water level (ft)}
 \item{smp_id}{factor, SMP ID}
 \item{ow_suffix}{factor, OW Suffix}
 \item{rainfall_in}{num, rainfall at each time step (in)}
 \item{gagename}{num, rain gage ID}
 \item{event}{num, rain event}
 
 }
}
\usage{
obs_250_all
}
\description{
A dataset containing water level, rain, and SMP data. \code{obs_250_fill} is joined with SMP stats.
}
\keyword{datasets}
