% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{obs_250_fill}
\alias{obs_250_fill}
\title{obs_250_fill}
\format{
a dataframe with 67820 rows and 14 variables
\describe{
 \item{dtime_est}{POSIXct datetime, YYYY-MM-DD HH:MM:SS}
 \item{level_ft}{num, water level (ft)}
 \item{smp_id}{factor, SMP ID}
 \item{ow_suffix}{factor, OW Suffix}
 \item{rainfall_in}{num, rainfall at each time step (in)}
 \item{gagename}{num, rain gage ID}
 \item{event}{num, rain event}
 \item{storage_depth_ft}{num, system storage depth (ft)}
 \item{storage_vol_ft3}{num, system storage volume (cf)}
 \item{infil_footprint_ft2}{num, system infiltration footprint (sf)}
 \item{dcia_ft2}{num, directly connected impervious area (sf)}
 \item{orifice_height_ft}{num, height of the slow release orifice relative to the bottom of stone storage (ft)}
 \item{orifice_diam_in}{num, orifice diameter (in)}
 \item{orifice_vol_cf}{num, volume leaving the orifice during the timestep (cf)}
}
}
\usage{
obs_250_fill
}
\description{
A dataset containing the data of \code{obs_250_all} joined with select SMP stats from \code{smp_stats}.
}
\keyword{datasets}
