% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_continuous_functions.R
\name{obsInfilRate_inhr}
\alias{obsInfilRate_inhr}
\title{Observed Infiltration Rate}
\usage{
obsInfilRate_inhr(event, dtime_est, rainfall_in, infil_footprint_ft2,
  dcia_ft2, orifice_height_ft = NA, orifice_diam_in = NA,
  storage_depth_ft, storage_vol_ft3, waterlevel_ft,
  discharge_coeff = 0.62)
}
\arguments{
\item{event}{Rainfall event ID (grouping variable)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}

\item{infil_footprint_ft2}{Total area of the system that is open to infiltration (sf)}

\item{dcia_ft2}{Directly connected impervious area (sf)}

\item{orifice_height_ft}{Orifice height (ft)}

\item{orifice_diam_in}{Orifice diameter (in)}

\item{storage_depth_ft}{Maximum storage depth of system (ft)}

\item{storage_vol_ft3}{Maximum storage volume (pore space) of system, in cubic feet}

\item{waterlevel_ft}{Observed water level data (ft)}

\item{discharge_coeff}{Orifice discharge coefficient}
}
\value{
Output is estimated infiltration rate (in/hr)
}
\description{
Estimated infiltration rate based on observed data
}
\examples{
obs_250_fill \%>\%
filter(is.na(event) == FALSE) \%>\%
  group_by(event) \%>\%
  arrange(dtime_est)\%>\%
  summarize( #Calculate performance metrics
    #Observed infiltration rate
    Infiltration_Rate_inhr = obsInfilRate_inhr(event, dtime_est,
                                               rainfall_in,
                                               infil_footprint_ft2,
                                               dcia_ft2,
                                               storage_depth_ft = storage_depth_ft,
                                               storage_vol_ft3 = storage_vol_ft3,
                                               orifice_diam_in,
                                               orifice_height_ft,
                                               waterlevel_ft = level_ft))

}
\seealso{
\code{\link{obsRecessionRate_inhr}}, \code{\link{obsOrificeOutVol_cf}}
}
