% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_continuous_functions.R
\name{obsOrificeOutVol_cf}
\alias{obsOrificeOutVol_cf}
\title{Observed Orifice Outflow Volume}
\usage{
obsOrificeOutVol_cf(dtime_est, waterlevel_ft, orifice_height_ft,
  orifice_diam_in, discharge_coeff = 0.62)
}
\arguments{
\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{waterlevel_ft}{Observed water level data (ft)}

\item{orifice_height_ft}{Orifice height (ft)}

\item{orifice_diam_in}{Orifice diameter (in)}

\item{discharge_coeff}{Orifice discharge coefficient}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}
}
\value{
Output is total observed orifice outflow volume (cf)
}
\description{
Total observed orifice outflow volume (cf)
}
\seealso{
\code{\link{obsInfilRate_inhr}}, \code{\link{obsRecessionRate_inhr}}
}
