% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_continuous_functions.R
\name{obsRecessionRate_inhr}
\alias{obsRecessionRate_inhr}
\title{Observed Recession Rate}
\usage{
obsRecessionRate_inhr(event, dtime_est, rainfall_in, waterlevel_ft)
}
\arguments{
\item{event}{Rainfall event ID (grouping variable)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}

\item{waterlevel_ft}{Observed water level data (ft)}
}
\value{
Estimated recession rate (in/hr)
}
\description{
Estimated recession rate based on observed data
}
\seealso{
\code{\link{obsInfilRate_inhr}}, \code{\link{obsOrificeOutVol_cf}}
}
