% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Rainfall}
\format{An object of class \code{data.frame} with 105 rows and 3 columns.}
\usage{
\code{example_rain_gage}
}
\description{
A dataset containing datetimes, rainfall in inches, and rainfall gage id.
}
\keyword{datasets}
