% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{simPlotNested}
\alias{simPlotNested}
\title{Simulated Water Levels Plot}
\usage{
simPlotNested(event, df, raingage_colname_or_value,
  system_colname_or_value, datetime_colname, rainfall_colname,
  storage_depth_colname_or_value_ft, storage_vol_colname_or_value_ft3,
  simwaterlevel_colname, orifice_show = FALSE,
  orifice_height_colname_or_value_ft = NULL)
}
\arguments{
\item{event}{Rainfall Event ID (grouping variable)}

\item{raingage_colname_or_value}{Raingage ID, or name of the column containing the Raingage ID information}

\item{datetime_colname}{Name of the column containing the datetime data}

\item{rainfall_colname}{Name of the column heading containing the rainfall data, in inches}

\item{storage_depth_colname_or_value_ft}{Maximum storage depth, in feet, or name of the column containing the 
system depth information}

\item{storage_vol_colname_or_value_ft3}{Maximum storage volume (pore space), in cubic feet, or name of the column
containing the system volume information}

\item{simwaterlevel_colname}{Name of the column containing the simulated water level data}

\item{orifice_show}{TRUE if user wants to include the orifice height as dashed line on plot (optional)}

\item{event}{df Dataframe passed from nested group in dplyr. The user must specify the column names 
in the function arguments, though the data is already stored in the dataframe}

\item{system_colname_or_valuee}{SMP ID, or name of the column containing the system ID}

\item{obswaterlevel_colname}{Name of the column containing the observed water level data}

\item{sim_event_colname}{Name of the column containing the rainfall event ID used for simulated timeseries generation}

\item{orifice_height_ft}{Orifice height, in feet, OR column name containing data}
}
\value{
Plot of simulated depth and volume with rainfall
}
\description{
Plot of simulated depth and volume with rainfall
}
