% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{simulated_data}
\format{
a dataframe with 1309 rows and 6 columns
  \describe{
  \item{dtime_edt}{POSIXct datetime, YYYY-MM-DD HH:MM:SS}
  \item{rainfall_in}{num, rainfall total in the 15 minutes preceding the corresponding datetime (in)}
  \item{event}{int, rainfall event uid}
  \item{simulated_depth_ft}{simulated water level in observation well (ft)}
  \item{simulated_vol_ft3}{simulated volume in the system (cf)}
  \item{simulated_orifice_vol_ft3}{simulated volume leaving the orifice (cf)}
  }

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1309 rows and 6 columns.
}
\usage{
simulated_data

testing
}
\description{
An example dataframe of simulated draindown data

An example list of observed data for system 1265-7. Used for testing function performance.
}
\keyword{datasets}
