% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_simulation_functions.R
\name{simulation.stats}
\alias{simulation.stats}
\alias{marsOvertoppingCheck_bool}
\alias{marsPeakStorage_percent}
\alias{marsPeakReleaseRate_cfs}
\alias{marsDraindown_hr}
\title{Simulation Stats}
\usage{
marsOvertoppingCheck_bool(waterlevel_ft, storage_depth_ft)

marsPeakStorage_percent(waterlevel_ft, storage_depth_ft)

marsPeakReleaseRate_cfs(dtime_est, orifice_outflow_ft3)

marsDraindown_hr(dtime_est, rainfall_in, waterlevel_ft)
}
\arguments{
\item{waterlevel_ft}{Water Level (ft)}

\item{storage_depth_ft}{Total storage Depth (ft)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{orifice_outflow_ft3}{Orifice outflow volume (cf)}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}
}
\value{
\describe{
     \item{\code{marsOvertoppingCheck_bool}}{Output is true or false based on overtopping evaluation}
}

\describe{
     \item{\code{marsPeakStorage_percent}}{Output is a percentage of peak storage filled, by depth}
}

\describe{
     \item{\code{marsPeakReleaseRate_cfs}}{Output is peak orifice release rate (cfs)}
}

\describe{
     \item{\code{marsDraindown_hr}}{Output is Calculated Draindown time (hr)}
}
}
\description{
Summarize and Evaluate System Performance
}
\examples{

simulation_summary <- simulated_data \%>\%
  dplyr::group_by(event) \%>\%
  dplyr::summarize(startdate = min(dtime_est), #add start date of event to summary table,
                  
   #1. Overtopping check
   overtopping = marsOvertoppingCheck_bool(Simulated_depth_ft,smp_stats$storage_depth_ft[7]),

   #2. Simulated storage utilization
   peakUtilization = marsPeakStorage_percent(Simulated_depth_ft,smp_stats$storage_depth_ft[7]),

   #3. Peak release rate
   peakReleaseRate_cfs = marsPeakReleaseRate_cfs(dtime_est, Simulated_orifice_vol_ft3),

   #4. Total orifice outflow volume (rounded for table format)
   orifice_volume_ft3 = round(sum(Simulated_orifice_vol_ft3),0),

   #5. Draindown time
   draindown_time_hr = marsDraindown_hr(dtime_est, rainfall_in, Simulated_depth_ft))
   
}
