% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_simulation_functions.R
\name{simWaterLevel_ft}
\alias{simWaterLevel_ft}
\title{Simulated Water Level}
\usage{
simWaterLevel_ft(dtime_est, rainfall_in, event, infil_footprint_ft2,
  dcia_ft2, orifice_height_ft = NA, orifice_diam_in = NA,
  storage_depth_ft, storage_vol_ft3, infil_rate_inhr,
  initial_water_level_ft = 0, runoff_coeff = 1,
  discharge_coeff = 0.62)
}
\arguments{
\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}

\item{event}{A vector of Event IDs}

\item{infil_footprint_ft2}{Total area of the system that is open to infiltration (sf)}

\item{dcia_ft2}{Directly connected impervious area (sf)}

\item{orifice_height_ft}{Orifice height (ft) (NA if no orifice outlet)}

\item{orifice_diam_in}{Orifice diameter (in) (NA if no orifice outlet)}

\item{storage_depth_ft}{Maximum storage depth (ft)}

\item{storage_vol_ft3}{Maximum storage volume (pore space) (cf)}

\item{infil_rate_inhr}{System design infiltration rate (in/hr)}

\item{initial_water_level_ft}{Initial water Level (ft) (Default = 0)}

\item{runoff_coeff}{Rational method coefficient (Default = 1)}

\item{discharge_coeff}{Orifice discharge coefficient (Defauly = 0.62)}
}
\value{
Output is a dataframe with the following columns: dtime_est, rainfall_in, event, Simulated_depth_ft, Simulated_vol_ft3, Simulated_orifice_vol_ft3
}
\description{
Simulates water level in subsurface stormwater infiltration system with underdrain
}
