% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_water_level_functions.R
\name{smoothing_marsInfiltrationRate_inhr}
\alias{smoothing_marsInfiltrationRate_inhr}
\title{Smoothed Infiltration Rate (Testing)}
\usage{
smoothing_marsInfiltrationRate_inhr(event, dtime_est, rainfall_in,
  dcia_ft2, orifice_height_ft = NA, orifice_diam_in = NA,
  storage_depth_ft, storage_vol_ft3, waterlevel_ft,
  discharge_coeff = 0.62)
}
\arguments{
\item{event}{Rainfall event ID (grouping variable)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}

\item{dcia_ft2}{Directly connected impervious area (sf)}

\item{orifice_height_ft}{Orifice height (ft)}

\item{orifice_diam_in}{Orifice diameter (in)}

\item{storage_depth_ft}{Maximum storage depth of system (ft)}

\item{storage_vol_ft3}{Maximum storage volume (pore space) of system, in cubic feet}

\item{waterlevel_ft}{Observed water level data (ft)}

\item{discharge_coeff}{Orifice discharge coefficient}
}
\value{
Output is estimated infiltration rate (in/hr). These outputs are codes for the following messages:
 \describe{
       \item{\code{-900}}{Event does not include observation data that approximately equals 5 or 7 in. water depth}
       \item{\code{-910}}{Code captures rising limb in event.}
       \item{\code{-920}}{Rainfall occurs during recession period between 7 in. and 5in. }
 }
 If
}
\description{
Testing function. Assure that the orifice outflow does not exceed delta V at each timestep
Smooth the data to 45 minutes to remove noise
}
\details{
Estimated infiltration rate based on observed data
}
\seealso{
\code{\link{marsUnderdrainOutflow_cf}}
}
