% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{smp_stats}
\alias{smp_stats}
\title{smp_stats}
\format{
a dataframe with 8 rows and 11 columns
#' \describe{
 \item{smp_id}{string, SMP ID}
 \item{ow_suffix}{factor, OW Suffix}
 \item{smp_footprint_ft2}{num, SMP surface area footprint (sf)}
 \item{infil_footprint_ft2}{num, SMP infiltration surface area footprint (sf)}
 \item{dcia_ft2}{num, directly connected impervious area (sf)}
 \item{orifice_height_ft}{num, height of the slow release orifice relative to the bottom of stone storage (ft)}
 \item{orifice_diam_in}{num, orifice diameter (in)}
 \item{storage_depth_ft}{num, system storage depth (ft)}
 \item{sump_depth_ft}{num, observation well sump depth (ft)}
 \item{storage_vol_ft3}{num, system storage volume (cf)}
 \item{infil_rate_inhr}{infiltration rate (in/hr)}
}
}
\usage{
smp_stats
}
\description{
A dataset containing summary statistics for 8 SMPs.
}
\keyword{datasets}
