% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{snapshot_250}
\alias{snapshot_250}
\title{SMP Snapshot}
\format{a dataframe with 1 row and 16 columns}
\usage{
\code{snapshot_250}
}
\description{
A dataset containing stats like footprint and storage depth for SMP 250-1-1
}
\keyword{datasets}
