% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{data}
\name{summary_250}
\alias{summary_250}
\title{SMP Summary}
\format{a dataframe with 27 rows and 11 columns}
\usage{
\code{summary_250}
}
\description{
A dataset containing summary statistics for SMP 250-1-1 from 2018-01-01 to 2018-03-01.
}
\keyword{datasets}
