% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{list}
\name{sys1265_data}
\alias{sys1265_data}
\title{sys1265_data}
\format{
An object of class \code{list} of length 3.
}
\usage{
testing
}
\description{
An example list of observed data for system 1265-7. Used for testing function performance.
}
\keyword{datasets}
