% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_documentation.R
\docType{list}
\name{sys1265_mets}
\alias{sys1265_mets}
\title{sys1265_mets}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
testing
}
\description{
An example list of metrics for system 1265-7. Used for testing function performance.
}
\keyword{datasets}
