% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_functions.R
\name{yday_decimal}
\alias{yday_decimal}
\title{Fetch a decimal day from a datetime}
\usage{
yday_decimal(dtime_est)
}
\arguments{
\item{dtime_est}{POSIXct, format: "POSIXct, format: "YYYY-MM-DD HH:MM:SS""}
}
\value{
Output with be a day with a decimal
}
\description{
Return the day of the year, with hours and seconds as the decimal
}
