% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{baroRasterPlot}
\alias{baroRasterPlot}
\title{Barometric Pressure Raster Plot}
\usage{
baroRasterPlot(baro)
}
\arguments{
\item{baro}{a dataframe with columns: \code{smp_id, baro_psi, day, year}}
}
\value{
p, a ggplot2 plot
}
\description{
Create a raster plot of barometric pressurs from each sensor for each day
}
\examples{
marsSampleBaro_plot \%<>\% dplyr::mutate("day" = yday_decimal(marsSampleBaro_plot$dtime_est),
                               "year" = lubridate::year(marsSampleBaro_plot$dtime_est))
baroRasterPlot(marsSampleBaro_plot)

}
