% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{Check if a date is the proper yyyy-mm-dd format}
\usage{
check_date(date)
}
\arguments{
\item{date}{Date value to check}

\item{tz}{Timezone character value}
}
\value{
The date with yyyy-mm-dd format or NULL if wrong format with warning
}
\description{
Check if a date is the proper yyyy-mm-dd format
}
