% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_functions.R
\name{detectEvents}
\alias{detectEvents}
\title{Identify individual rainfall events}
\usage{
detectEvents(dtime_edt, rainfall_in, iet_hr = 6, mindepth_in = 0.1)
}
\arguments{
\item{dtime_edt}{vector, POSIXct date times}

\item{rainfall_in}{vector, num, of rainfall depths corresponding to \code{dtime_edt}, in inches}

\item{iet_hr}{num, Interevent time, in hours. The default is 6 hours.}

\item{mindepth_in}{num, minimum depth of a given event, in inches. The default is 0.10 inches.}
}
\value{
Output will be a vector of integers corresponding to \code{dtime_edt} and representing
  the event ID for each time step.
}
\description{
Return a dataset of rainfall event IDs for a time period
}
