% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_smp_lookup.R
\name{lookupPrivateSMPs}
\alias{lookupPrivateSMPs}
\title{Return query results of private SMPs}
\usage{
lookupPrivateSMPs(con, tracking_numbers)
}
\arguments{
\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}

\item{tracking_number}{chr, tracking number for the requested SMP}
}
\value{
Output will be a dateframe. For tracking numbers were found in database, output will be private SMP
  "tracking_number" (chr), "project name" (chr), "smp_id" (int), and "plan_label" (chr).
  If a tracking number is not found, that row will include an error message
  in the "project name" column, and NAs in "smp_id" and "plan_label".
}
\description{
Returns private SMP tracking number, project name, SMP ID, and plan label.
}
