% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsCombinedPlot}
\alias{marsCombinedPlot}
\title{Plot hyetograph and water level plot on the same chart}
\usage{
marsCombinedPlot(
  event,
  structure_name,
  obs_datetime,
  obs_level_ft,
  sim_datetime = NA,
  sim_level_ft = NA,
  storage_depth_ft,
  orifice_show = FALSE,
  orifice_height_ft = NULL,
  rainfall_datetime,
  rainfall_in,
  obs_RSPU = NA,
  sim_RSPU = NA,
  obs_infil = NA,
  sim_infil = NA,
  obs_draindown = NA,
  sim_draindown = NA,
  obs_overtopping = NA,
  sim_overtopping = NA
)
}
\arguments{
\item{event}{chr, rain gage event UID}

\item{structure_name}{chr, SMP ID and OW Suffix}

\item{obs_datetime}{vector, POSIXct datetimes corresponding to \code{obs_level_ft}}

\item{obs_level_ft}{vector, water level data (ft), corresponding to \code{obs_datetime}}

\item{sim_datetime}{vector, POSIXct datetimes corresponding to \code{sim_level_ft} (optional)}

\item{sim_level_ft}{vector, water level data (ft), corresponding to \code{sim_datetime} (optional)}

\item{storage_depth_ft}{num, maximum storage depth of system (ft)}

\item{orifice_show}{TRUE if user wants to include the orifice height as dashed line on plot (optional)}

\item{orifice_height_ft}{Orifice height, in feet (optional)}

\item{rainfall_datetime}{vector, POSIXct datetimes corresponding to \code{rainfall_in}}

\item{rainfall_in}{vector, num, rainfall in inches corresponding to \code{rainfall_datetime}}

\item{obs_RSPU}{num, Metric: Observed relative percentage of storage used, see \code{marsPeakStorage_percent} (optional)}

\item{sim_RSPU}{num, Metric: Simulated relative percentage of storage used, see \code{marsPeakStorage_percent} (optional)}

\item{obs_infil}{num, Metric: Observed infiltration rate in inches per hour, see \code{marsInfiltrationRate_inhr} (optional)}

\item{sim_infil}{num, Metric: Simulated infiltration rate in inches per hour, see \code{marsInfiltrationRate_inhr} (optional)}

\item{obs_draindown}{num, Metric: Observed draindown time in hours, see \code{marsDraindown_hr} (optional)}

\item{sim_draindown}{num, Metric: Simulated draindown time in hours, see \code{marsDraindown_hr} (optional)}

\item{obs_overtopping}{bool, Metric: Observed overtopping boolean, see \code{marsOvertoppingCheck_bool} (optional)}

\item{sim_overtopping}{bool, Metric: Simulated overtopping boolean, see \code{marsOvertoppingCheck_bool} (optional)}

\item{metrics_show}{bool, Default FALSE. TRUE if user wants to include a table of metrics on the plot (optional)}
}
\value{
Output will be a gridExtra object of the two plots
}
\description{
Return hyetograph and observed and simulated (optional) water level plot for the same rain event on the same chart
}
\seealso{
\code{\link{marsRainfallPlot}}, \code{\link{marsWaterLevelPlot}}
}
