% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsCombinedPlot}
\alias{marsCombinedPlot}
\title{Plot hyetograph and water level plot on the same chart}
\usage{
marsCombinedPlot(event, structure_name, level_datetime, storage_depth_ft,
  level_ft, rainfall_datetime, rainfall_in, raingage)
}
\arguments{
\item{event}{chr, rain gage event UID}

\item{structure_name}{chr, SMP ID and OW Suffix}

\item{level_datetime}{vector, POSIXct datetimes corresponding to \code{level_ft}}

\item{storage_depth_ft}{num, maximum storage depth of system (ft)}

\item{level_ft}{vector, water level data (ft), corresponding to \code{level_datetime}}

\item{rainfall_datetime}{vector, POSIXct datetimes corresponding to \code{rainfall_in}}

\item{rainfall_in}{vector, num, rainfall in inches corresponding to \code{rainfall_datetime}}

\item{raingage}{chr, Label for the hyetograph for what rain gage the data came from}
}
\value{
Output will be a gridExtra object of the two plots
}
\description{
Return hyetograph and water level plot for the same rain event on the same chart
}
\seealso{
\code{\link{marsRainfallPlot}}, \code{\link{marsWaterLevelPlot}}
}
