% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsFetchBaroData}
\alias{marsFetchBaroData}
\title{Fetch barometric pressure data for a target SMP, date range, and interval}
\usage{
marsFetchBaroData(
  con,
  target_id,
  start_date,
  end_date,
  data_interval = c("5 mins", "15 mins")
)
}
\arguments{
\item{con}{An ODBC connection to the MARS Analysis database returned by odbc::dbConnect}

\item{target_id}{chr, single SMP ID where the user has requested data}

\item{start_date}{string or POSIXct, format: "YYYY-MM-DD", start of data request range}

\item{end_date}{string or POSIXct, format: "YYYY-MM-DD", end of data request range}

\item{data_interval}{chr, \code{"5 mins"} or \code{"15 mins"}, interval at which baro data will be returned.}
}
\value{
Output will be a dataframe with four columns: 
  
    \item{dtime}{POSIXct, format: "YYYY-MM-DD HH:MM:SS" in America/New_York}
    \item{baro_psi}{num, barometric pressure in psi}
    \item{smp_id}{chr, SMP ID for each baro}
    
    If there are fewer than five baros to interprolate from, based on \code{\link{marsInterpolateBaro}}
}
\description{
Returns a data frame with datetime, barometric pressure, smp id, and number of neighbors
  interpolated from to collect the data.
}
\seealso{
\code{\link{marsInterpolateBaro}}
}
