% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsFetchRainEventData}
\alias{marsFetchRainEventData}
\title{Fetch rain event data for an SMP}
\usage{
marsFetchRainEventData(con, target_id, start_date, end_date)
}
\arguments{
\item{con}{An ODBC connection to the MARS Analysis database returned by odbc::dbConnect}

\item{target_id}{vector of chr, SMP ID, where the user has requested data}

\item{start_date}{POSIXct, format: "YYYY-MM-DD", start of data request range}

\item{end_date}{POSIXct, format: "YYYY-MM-DD", end of data request range}
}
\value{
Output will be a dataframe with the following columns: 

    \item{rainfall_gage_event_uid}{int}
    \item{gage_uid}{int, rain gage number}
    \item{eventdatastart_edt}{POSIXct datetime}
    \item{eventdataend_edt}{POSIXct datetime}
    \item{eventduration_hr}{num, duration of event}
    \item{eventpeakintensity_inhr}{num, peak intensity of rain event} 
    \item{eventavgintensity_inhr}{num, average intensity of rain event} 
    \item{eventdepth_in}{num, average intensity of rain event}
}
\description{
Returns a data frame with rain event data from the rain gage closest to the request SMP
}
\seealso{
\code{\link{marsFetchRainGageData}}, \code{\link{marsFetchLevelData}}, \code{\link{marsFetchMonitoringData}}
}
