% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_water_level_functions.R
\name{marsInfiltrationRate_inhr}
\alias{marsInfiltrationRate_inhr}
\title{Infiltration Rate}
\usage{
marsInfiltrationRate_inhr(
  event,
  dtime_est,
  rainfall_in,
  dcia_ft2,
  orifice_height_ft = NA,
  orifice_diam_in = NA,
  storage_depth_ft,
  storage_vol_ft3,
  waterlevel_ft,
  depth_in = 6,
  discharge_coeff = 0.62
)
}
\arguments{
\item{event}{Rainfall event ID (grouping variable)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}

\item{dcia_ft2}{Directly connected impervious area (sf)}

\item{orifice_height_ft}{Orifice height (ft)}

\item{orifice_diam_in}{Orifice diameter (in)}

\item{storage_depth_ft}{Maximum storage depth of system (ft)}

\item{storage_vol_ft3}{Maximum storage volume (pore space) of system, in cubic feet}

\item{waterlevel_ft}{Observed water level data (ft)}

\item{depth_in}{Depth at which to calculate infiltration rate (default 6in, which creates a range of 7in to 5in)}

\item{discharge_coeff}{Orifice discharge coefficient (default 0.62)}
}
\value{
Output is estimated infiltration rate (in/hr). These outputs are codes for the following messages: 
 \describe{
       \item{\code{-900}}{Event does not include observation data that approximately equals specified water depth}
       \item{\code{-910}}{Code captures rising limb in event.}
       \item{\code{-920}}{Rainfall occurs during recession period in specified depth range.}
       \item{\code{-930}}{Infiltration rate is negligible; calculated infiltration rate is less than 0.1in/hr. }
 }
}
\description{
Estimated infiltration rate based on observed data
}
\examples{
obs_250_fill \%>\%
  dplyr::filter(is.na(event) == FALSE) \%>\%
  dplyr::group_by(event) \%>\%
  dplyr::arrange(dtime_est)\%>\%
  dplyr::summarize( #Calculate performance metrics
    #Observed infiltration rate
    Infiltration_Rate_inhr = marsInfiltrationRate_inhr(event, dtime_est,
                                               rainfall_in,
                                               dcia_ft2,
                                               storage_depth_ft = storage_depth_ft,
                                               storage_vol_ft3 = storage_vol_ft3,
                                               orifice_diam_in,
                                               orifice_height_ft,
                                               waterlevel_ft = level_ft))

}
\seealso{
\code{\link{marsUnderdrainOutflow_cf}}
}
