% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsInterpolateBaro}
\alias{marsInterpolateBaro}
\title{Interpolate barometric pressure with inverse distance weighting}
\usage{
marsInterpolateBaro(baro_psi, smp_id, weight, target_id)
}
\arguments{
\item{baro_psi}{vector, num, barometric pressures measured at the same timestamp}

\item{smp_id}{vector, chr, SMP IDs where the measurements took place}

\item{target_id}{chr, single SMP ID where the user has requested data}

\item{weights}{vector, num, of inverse distances weights for each baro, calculated by \code{\link{marsFetchBaroData}}}
}
\value{
Output will be a single barometric pressure reading.
  If there are 4 or greater baros with data,
  the reading will be an inverse distance-weighted
  interpolation of those readings.
  If there are fewer than 4 readings, return \code{NA}.
}
\description{
Returns an interpolated barometric pressure reading or \code{NA}
}
\examples{

 data(marsSampleBaro)

 marsInterpolateBaro(
   baro_psi = marsSampleBaro[[1]]$baro_psi, 
   smp_id = marsSampleBaro[[1]]$smp_id, 
   weight = marsSampleBaro[[1]]$weight, 
   target_id = marsSampleBaro[[2]]
   )

}
\seealso{
\itemize{
     \code{\link{marsFetchBaroData}},
     data: \code{\link{marsSampleBaro}}
 }
}
