% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{marsSampleBaro_plot}
\alias{marsSampleBaro_plot}
\title{Sample Baro Data for Plotting}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2890 rows and 5 columns.}
\usage{
\code{marsBaroRasterPlot}
}
\description{
A dataset containing datetime, smp_id, barometric pressure, number of neighbors, and weight
}
\keyword{datasets}
