% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsWaterLevelPlot}
\alias{marsWaterLevelPlot}
\title{Observed and Simulated Water Level Plot}
\usage{
marsWaterLevelPlot(
  event,
  structure_name,
  storage_depth_ft,
  obs_datetime,
  obs_level_ft,
  sim_datetime = NA,
  sim_level_ft = NA,
  orifice_show = FALSE,
  orifice_height_ft = NULL,
  snapshot = NA,
  metrics_show = FALSE,
  obs_RSPU = NA,
  sim_RSPU = NA,
  obs_infil = NA,
  sim_infil = NA,
  obs_draindown = NA,
  sim_draindown = NA,
  obs_overtopping = NA,
  sim_overtopping = NA
)
}
\arguments{
\item{event}{Rainfall gage event ID}

\item{structure_name}{SMP ID and OW Suffix}

\item{storage_depth_ft}{Maximum storage depth of system (ft)}

\item{obs_datetime}{Vector of POSIXct datetimes for observed dataset}

\item{obs_level_ft}{Vector of water level data (ft), corresponding to \code{obs_datetime}}

\item{sim_datetime}{Vector of POSIXct datetimes for simulated dataset (optional)}

\item{sim_level_ft}{Vector of water level data (ft), corresponding to \code{sim_datetime} (optional)}

\item{orifice_show}{TRUE if user wants to include the orifice height as dashed line on plot (optional)}

\item{orifice_height_ft}{Orifice height, in feet (optional)}

\item{metrics_show}{Bool, Default FALSE. TRUE if user wants to include a table of metrics on the plot (optional)}

\item{obs_RSPU}{Num, Metric: Observed relative percentage of storage used, see \code{marsPeakStorage_percent} (optional)}

\item{sim_RSPU}{Num, Metric: Simulated relative percentage of storage used, see \code{marsPeakStorage_percent} (optional)}

\item{obs_infil}{Num, Metric: Observed infiltration rate in inches per hour, see \code{marsInfiltrationRate_inhr} (optional)}

\item{sim_infil}{Num, Metric: Simulated infiltration rate in inches per hour, see \code{marsInfiltrationRate_inhr} (optional)}

\item{obs_draindown}{Num, Metric: Observed draindown time in hours, see \code{marsDraindown_hr} (optional)}

\item{sim_draindown}{Num, Metric: Simulated draindown time in hours, see \code{marsDraindown_hr} (optional)}

\item{obs_overtopping}{Bool, Metric: Observed overtopping boolean, see \code{marsOvertoppingCheck_bool} (optional)}

\item{sim_overtopping}{Bool, Metric: Simulated overtopping boolean, see \code{marsOvertoppingCheck_bool} (optional)}
}
\value{
Output is a ggplot2 object of the water level plot.
}
\description{
Create a plot of observed and simulated (optional) water level
}
