% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_plotting_functions.R
\name{marsWaterLevelPlot}
\alias{marsWaterLevelPlot}
\title{Observed or Simulated Water Levels Plot}
\usage{
marsWaterLevelPlot(event, structure_name, datetime, storage_depth_ft,
  level_ft, orifice_show = FALSE, orifice_height_ft = NULL)
}
\arguments{
\item{event}{Rainfall gage event ID}

\item{structure_name}{SMP ID and OW Suffix}

\item{datetime}{Vector of POSIXct datetimes}

\item{storage_depth_ft}{Maximum storage depth of system (ft)}

\item{level_ft}{Vector of water level data (ft), corresponding to \code{datetime}}

\item{orifice_show}{TRUE if user wants to include the orifice height as dashed line on plot (optional)}

\item{orifice_height_ft}{Orifice height, in feet (optional)}
}
\value{
Output is a ggplot2 object of the water level plot.
}
\description{
Create a plot of water level depth
}
