% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsWriteDraindownData}
\alias{marsWriteDraindownData}
\title{Write Draindown Data to Database}
\usage{
marsWriteDraindownData(
  con,
  draindown_hr,
  draindown_assessment_lookup_uid,
  ow_uid,
  radar_event_uid,
  snapshot_uid,
  observed_simulated_lookup_uid
)
}
\arguments{
\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}

\item{draindown_hr}{vector, numeric, draindown time (hr)}

\item{ow_uid}{vector, numeric observation well UID}

\item{radar_event_uid}{vector, numeric}

\item{snapshot_uid}{vector, numeric}

\item{observed_simulated_lookup_uid}{vector, numeric, 1 if observed, 2 if simulated}
}
\value{
\code{TRUE} if the write is succesful, or an error message if unsuccessful
}
\description{
Receive vector of draindown data, calculated with \code{\link{marsDraindown_hr}},
and write to MARS Analysis Database performance_draindwown table
}
\examples{

marsWriteDraindownData(con,
  draindown_hr = summary_250$draindown_hr,
  ow_uid = summary_250$ow_uid,
  radar_event_uid = summary_250$rainfall_gage_event_uid, 
  snapshot_uid = summary_250$snapshot_uid,
  observed_simulated_lookup_uid = summary_250$observed_simulated_lookup_uid)

}
\seealso{
\code{\link{marsWritePercentStorageData}}, \code{\link{marsWriteOvertoppingData}},
}
