% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsWriteOvertoppingData}
\alias{marsWriteOvertoppingData}
\title{Write Overtopping Data to Database}
\usage{
marsWriteOvertoppingData(overtopping, observed_simulated_lookup_uid,
  ow_uid, rainfall_gage_event_uid, snapshot_uid)
}
\arguments{
\item{overtopping}{vector, logical, TRUE if water level reaches max storage depth}

\item{observed_simulated_lookup_uid}{vector, numeric, 1 if observed, 2 if simulated}

\item{ow_uid}{vector, numeric observation well UID}

\item{rainfall_gage_event_uid}{vector, numeric}

\item{snapshot_uid}{vector, numeric}

\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}
}
\value{
\code{TRUE} if the write is succesful, or an error message if unsuccessful
}
\description{
Receive vector of overtopping data, calculated with \code{\link{marsOvertoppingCheck_bool}},
and write to MARS Analysis Database performance_overtopping table
}
\seealso{
\code{\link[pwdgsi]{marsWriteSaturatedData}}, \code{\link{marsWritePercentStorageData}}
}
