% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_data_fetch_write_functions.R
\name{marsWritePercentStorageData}
\alias{marsWritePercentStorageData}
\title{Write Percent of Storaged Used Data to Database}
\usage{
marsWritePercentStorageData(con, percentstorageused_peak,
  percentstorageused_relative, ow_uid, rainfall_gage_event_uid,
  snapshot_uid, observed_simulated_lookup_uid)
}
\arguments{
\item{con}{Formal class PostgreSQL, a connection to the MARS Analysis database}

\item{ow_uid}{vector, numeric observation well UID}

\item{rainfall_gage_event_uid}{vector, numeric}

\item{snapshot_uid}{vector, numeric}

\item{observed_simulated_lookup_uid}{vector, numeric, 1 if observed, 2 if simulated}

\item{infiltration_rate_inhr}{vector, numeric, infiltration rate (in/hr)}

\item{recession_rate_inhr}{vector, numeric, recession rate (in/hr)}
}
\value{
\code{TRUE} if the write is succesful, or an error message if unsuccessful
}
\description{
Receive vectors of raw and relative percent storage data, calculated with \code{\link{marsPeakStorage_percent}},
gather data, and write to MARS Analysis performance_percentstorage table
}
\seealso{
\code{\link[pwdgsi]{marsWriteSaturatedData}}, \code{\link{marsWriteOvertoppingData}},  \code{\link{marsWriteDraindownData}}
}
