% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Rainfall}
\format{A dataframe with 105 rows and 3 variables
\describe{
  \item{dtime_edt}{POSIXct datetime, YYYY-MM-DD HH:MM:SS}
  \item{rainfall_in}{rainfall total in the 15 minutes preceding the corresponding datetime}
  \item{gage_uid}{rain gage ID}
}}
\usage{
\code{example_rain_gage}
}
\description{
A dataset containing datetimes, rainfall in inches, and rainfall gage id.
}
\keyword{datasets}
