% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_simulation_functions.R
\name{simulation.stats}
\alias{simulation.stats}
\alias{overtoppingCheck_bool}
\alias{peakStorUtil_percent}
\alias{orificePeakRelease_cfs}
\alias{draindown_hr}
\title{Simulation Stats}
\usage{
overtoppingCheck_bool(waterlevel_ft, storage_depth_ft)

peakStorUtil_percent(waterlevel_ft, storage_depth_ft)

orificePeakRelease_cfs(dtime_est, orifice_outflow_ft3)

draindown_hr(dtime_est, rainfall_in, waterlevel_ft)
}
\arguments{
\item{waterlevel_ft}{Water Level (ft)}

\item{storage_depth_ft}{Total storage Depth (ft)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}

\item{orifice_outflow_ft3}{Orifice outflow volume (cf)}

\item{rainfall_in}{Rainfall depths during periods corresponding to times in  dtime_est (in)}

\item{dtime_est}{A vector of POSIXct date times, in ascending order}
}
\value{
\describe{
     \item{\code{overtoppingCheck_bool}}{Output is true or false based on overtopping evaluation}
}

\describe{
     \item{\code{peakStorUtil_percent}}{Output is a percentage of peak storage filled, by depth}
}

\describe{
     \item{\code{orificePeakRelease_cfs}}{Output is peak orifice release rate (cfs)}
}

\describe{
     \item{\code{draindown_hr}}{Output is Calculated Draindown time (hr)}
}
}
\description{
Summarize and Evaluate System Performance
}
