% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_rainfall_functions.R
\name{storm}
\alias{storm}
\alias{marsStormDepth_in}
\alias{marsStormDuration_hr}
\alias{marsStormPeakIntensity_inhr}
\alias{marsStormAverageIntensity_inhr}
\title{Return storm attributes}
\usage{
marsStormDepth_in(rainfall_in)

marsStormDuration_hr(dtime_est)

marsStormPeakIntensity_inhr(dtime_est, rainfall_in)

marsStormAverageIntensity_inhr(dtime_est, rainfall_in)
}
\arguments{
\item{rainfall_in}{vector, num, rainfall depth in inches representing a single rain event}

\item{dtime_est}{vector, POSIXct date times representing a single rain event}
}
\value{
\describe{
       \item{\code{marsStormDepth_in}}{Output will be total rainfall depth for the event, in inches.}
}

\describe{
       \item{\code{marsStormDuration_hr}}{Output will be a double with the duration of the event, in hours.}
}

\describe{
       \item{\code{marsStormPeakIntensity_inhr}}{Output will be a number representing the event's peak intensity in inches/hour.}
}

\describe{
       \item{\code{marsStormAverageIntensity_inhr}}{Output will be a number representing the event's average intensity in inches/hour.}
}
}
\description{
Return storm depth, duration, average intensity, and peak intensity of an event processed using \code{\link{marsDetectEvents}}.
}
\examples{
rain_newevents <- marsSampleRain \%>\%  #use dplyr pipe to update dataframe
 group_by(gage_uid) \%>\% 
  arrange(dtime_est) \%>\% 
  mutate(event_id = marsDetectEvents(dtime_est, rainfall_in)) \%>\%
  group_by(gage_uid, event_id) \%>\%
  summarize(eventdatastart_edt = first(dtime_est),
            eventdataend_edt = last(dtime_est),
            eventduration_hr = marsStormDuration_hr(dtime_est),
            eventpeakintensity_inhr = marsStormPeakIntensity_inhr(dtime_est, rainfall_in),
            eventavgintensity_inhr = marsStormAverageIntensity_inhr(dtime_est, rainfall_in),
            eventdepth_in = marsStormDepth_in(rainfall_in))
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{arrange}},
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{summarise}},
 \code{\link[dplyr]{select}}
}
