% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_functions.R
\name{storm}
\alias{storm}
\alias{stormDepth_in}
\alias{stormDuration_hr}
\alias{stormPeakIntensity_inhr}
\alias{stormAvgIntensity_inhr}
\title{Return storm attributes}
\usage{
stormDepth_in(rainfall_in)

stormDuration_hr(dtime_edt)

stormPeakIntensity_inhr(dtime_edt, rainfall_in)

stormAvgIntensity_inhr(dtime_edt, rainfall_in)
}
\arguments{
\item{rainfall_in}{vector, num, rainfall depth in inches representing a single rain event}

\item{dtime_edt}{vector, POSIXct date times representing a single rain event}
}
\value{
\describe{
       \item{\code{\link{stormDepth_in}}}{Output will be total rainfall depth for the event, in inches.}
}

\describe{
       \item{\code{\link{stormDuration_hr}}}{Output will be a double with the duration of the event, in hours.}
}

\describe{
       \item{\code{\link{stormPeakIntensity_inhr}}}{Output will be a number representing the event's peak intensity in inches/hour.}
}

\describe{
       \item{\code{\link{stormAvgIntensity_inhr}}}{Output will be a number representing the event's average intensity in inches/hour.}
}
}
\description{
Return storm depth, duration, average intensity, and peak intensity of an event processed using \code{\link{detectEvents}}.
}
